requirements.txt
fastapi>=0.95.0,<0.104.0
uvicorn>=0.15.0,<0.23.0
openai>=1.0.0
httpx>=0.22.0,<0.25.0
pydantic>=1.9.0,<2.0.0
python-dotenv>=0.19.0
requests>=2.28.0 